Updated/modified data visualization chapters for tables and figures.
Added information about typesetting in LaTeX, including information on citation and referencing.

Added Make target to the Makefile.

Modified preamble of the files for the main report, and the others subdirectory for files specifying the colors that I can use for writing in LaTeX and the revision history of the file.